add get spots service, and try to draw it

+++ app/src/main/java/smarttraffic/smartparking/dataModels/SmartParkingSpot.java
@@ -11,16 +11,16 @@ public class SmartParkingSpot {
 
     private Integer id;
     private Integer in_lot;
-    private float p1_latitud;
-    private float p1_longitud;
-    private float p2_latitud;
-    private float p2_longitud;
-    private float p3_latitud;
-    private float p3_longitud;
-    private float p4_latitud;
-    private float p4_longitud;
-    private float p5_latitud;
-    private float p5_longitud;
+    private double p1_latitud;
+    private double p1_longitud;
+    private double p2_latitud;
+    private double p2_longitud;
+    private double p3_latitud;
+    private double p3_longitud;
+    private double p4_latitud;
+    private double p4_longitud;
+    private double p5_latitud;
+    private double p5_longitud;
     private String status;
     private Date created;
     private Date updated;
@@ -62,83 +62,83 @@ public class SmartParkingSpot {
         this.in_lot = in_lot;
     }
 
-    public float getP1_latitud() {
+    public double getP1_latitud() {
         return p1_latitud;
     }
 
-    public void setP1_latitud(float p1_latitud) {
+    public void setP1_latitud(double p1_latitud) {
         this.p1_latitud = p1_latitud;
     }
 
-    public float getP1_longitud() {
+    public double getP1_longitud() {
         return p1_longitud;
     }
 
-    public void setP1_longitud(float p1_longitud) {
+    public void setP1_longitud(double p1_longitud) {
         this.p1_longitud = p1_longitud;
     }
 
-    public float getP2_latitud() {
+    public double getP2_latitud() {
         return p2_latitud;
     }
 
-    public void setP2_latitud(float p2_latitud) {
+    public void setP2_latitud(double p2_latitud) {
         this.p2_latitud = p2_latitud;
     }
 
-    public float getP2_longitud() {
+    public double getP2_longitud() {
         return p2_longitud;
     }
 
-    public void setP2_longitud(float p2_longitud) {
+    public void setP2_longitud(double p2_longitud) {
         this.p2_longitud = p2_longitud;
     }
 
-    public float getP3_latitud() {
+    public double getP3_latitud() {
         return p3_latitud;
     }
 
-    public void setP3_latitud(float p3_latitud) {
+    public void setP3_latitud(double p3_latitud) {
         this.p3_latitud = p3_latitud;
     }
 
-    public float getP3_longitud() {
+    public double getP3_longitud() {
         return p3_longitud;
     }
 
-    public void setP3_longitud(float p3_longitud) {
+    public void setP3_longitud(double p3_longitud) {
         this.p3_longitud = p3_longitud;
     }
 
-    public float getP4_latitud() {
+    public double getP4_latitud() {
         return p4_latitud;
     }
 
-    public void setP4_latitud(float p4_latitud) {
+    public void setP4_latitud(double p4_latitud) {
         this.p4_latitud = p4_latitud;
     }
 
-    public float getP4_longitud() {
+    public double getP4_longitud() {
         return p4_longitud;
     }
 
-    public void setP4_longitud(float p4_longitud) {
+    public void setP4_longitud(double p4_longitud) {
         this.p4_longitud = p4_longitud;
     }
 
-    public float getP5_latitud() {
+    public double getP5_latitud() {
         return p5_latitud;
     }
 
-    public void setP5_latitud(float p5_latitud) {
+    public void setP5_latitud(double p5_latitud) {
         this.p5_latitud = p5_latitud;
     }
 
-    public float getP5_longitud() {
+    public double getP5_longitud() {
         return p5_longitud;
     }
 
-    public void setP5_longitud(float p5_longitud) {
+    public void setP5_longitud(double p5_longitud) {
         this.p5_longitud = p5_longitud;
     }
